fix: return type inconsistency in `createBlurrableHandler`
We used a `void` as the return type for the handler, yet at the same time we were returning the result from the handler itself. Sometimes we actually might return a value (though not common); for this reason, this commit updates the TypeScript definitions to return the same type as the handler.