fix: ensure "None" property unique features aren't overwritten
The "None" property has unique handling, which compares the entire value to `0`, rather than testing if the `0` value is enabled within the bitwise (as done for other values), and also doesn't have a setter. A missing `continue` statement caused the flow of code to fall through and execute the default property definer, which overwrites the unique implementation.

Add this edge-case as a unit-test respectively.

Co-Authored-By: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>