fix: extract TimePicker component and use bootstrap event handlers
Extracting the TimePicker to its own component makes sense due to the massive logic with jQuery there.

Another issue is that we registered the `onChange` event for the input and updated the state on every keystroke, which is not compatible with the bootstrap's timepicker we were using (per Waze's implementation). It caused issues when trying to change the value of an already populated time picker field.
To solve this bug, in the extracted component we register to the relevant bootstrap events that are fired after the user has done typing the new time and boostrap has processed the raw value.