refactor!: integrate rollup-plugin-userscript for adding the metadata (#103)
Integrates a new rollup plugin that generates a userscript's metadata block automatically during build. It is configured to read the script's name from `package.json` and parse it (either by using the `displayName` property from `package.json`, or by transforming the `name` property to Space Case), to automatically detect `grant` statements, with all Waze domains to match (including internal ones), importing WME SDK+, and allowing a connection to Crowdin.


BREAKING CHANGE:
In addition, due to the minimal use of the development server, and complexities with keeping the static metadata file for it in sync with the automated metadata block, the development server is removed altogether from the script. Using a development server is no longer possible.